Fold Drupal support into QC report and Pub Preview
